refactor:FilePreview 파일 이름 및 가져오기 변경
